PLFM-1003

+++ lib/jdomodels/src/test/java/org/sagebionetworks/repo/model/dbo/persistence/DBOUserProfileTest.java
@@ -1,12 +1,20 @@
 package org.sagebionetworks.repo.model.dbo.persistence;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+import java.util.Date;
 
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.sagebionetworks.repo.model.DatastoreException;
+import org.sagebionetworks.repo.model.UserGroup;
+import org.sagebionetworks.repo.model.UserGroupDAO;
 import org.sagebionetworks.repo.model.dbo.DBOBasicDao;
+import org.sagebionetworks.repo.model.jdo.KeyFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.test.context.ContextConfiguration;
@@ -19,22 +27,52 @@ public class DBOUserProfileTest {
 	@Autowired
 	DBOBasicDao dboBasicDao;
 	
-	private long id = 1000;
+	@Autowired 
+	UserGroupDAO userGroupDAO;
+	
+	private static final String TEST_USER_NAME = "test-user";
+	
+	private UserGroup individualGroup = null;
+	
+	
+	@Before
+	public void setUp() throws Exception {
+		individualGroup = userGroupDAO.findGroup(TEST_USER_NAME, true);
+		if (individualGroup == null) {
+			individualGroup = new UserGroup();
+			individualGroup.setName(TEST_USER_NAME);
+			individualGroup.setIndividual(true);
+			individualGroup.setCreationDate(new Date());
+			individualGroup.setId(userGroupDAO.create(individualGroup));
+		}
+		deleteUserProfile();
+	}
+	
+	private void deleteUserProfile() throws DatastoreException {
+		if(dboBasicDao != null && individualGroup!=null){
+			MapSqlParameterSource params = new MapSqlParameterSource();
+			params.addValue("ownerId", individualGroup.getId());
+			dboBasicDao.deleteObjectById(DBOUserProfile.class, params);
+		}		
+	}
+		
 	
 	@After
-	public void after() throws DatastoreException{
-		if(dboBasicDao != null){
-			MapSqlParameterSource params = new MapSqlParameterSource();
-			params.addValue("id", id);
-			dboBasicDao.deleteObjectById(DBOUserProfile.class, params);
+	public void tearDown() throws Exception{
+		individualGroup = userGroupDAO.findGroup(TEST_USER_NAME, true);
+		if (individualGroup != null) {
+			// this will delete the user profile too
+			userGroupDAO.delete(individualGroup.getId());
 		}
 	}
+	
 	@Test
 	public void testCRUD() throws Exception{
 		// Create a new type
 		DBOUserProfile userProfile = new DBOUserProfile();
-		userProfile.setId(id);
-		userProfile.setUserName("foo@bar.bas");
+		userProfile.setOwnerId(KeyFactory.stringToKey(individualGroup.getId()));
+		userProfile.seteTag(10L);
+		userProfile.setProperties("My dog has fleas.".getBytes());
 		
 		// Create it
 		DBOUserProfile clone = dboBasicDao.createNew(userProfile);
@@ -42,11 +80,10 @@ public class DBOUserProfileTest {
 		assertEquals(userProfile, clone);
 		// Fetch it
 		MapSqlParameterSource params = new MapSqlParameterSource();
-		params.addValue("id", id);
+		params.addValue("ownerId", individualGroup.getId());
 		clone = dboBasicDao.getObjectById(DBOUserProfile.class, params);
 		assertNotNull(clone);
-		assertEquals(userProfile.getId(), clone.getId());
-		assertEquals(userProfile.getUserName(), clone.getUserName());
+		assertEquals(userProfile.getOwnerId(), clone.getOwnerId());
 		// Delete it
 		boolean result = dboBasicDao.deleteObjectById(DBOUserProfile.class,  params);
 		assertTrue("Failed to delete the type created", result);
